requirements.txt
streamlit
sqlalchemy
psycopg2-binary
pandas
requests
apscheduler

# Co-pilot libraries
PyMuPDF
python-docx
sentence-transformers
faiss-cpu
ctransformers
ddgs

# Phase 3 & 4 Dependencies
# Email functionality for RFQ system
sendgrid>=6.9.7
python-dotenv>=0.19.0

# Enhanced document generation
reportlab>=3.6.0

# Enhanced web scraping and data processing
beautifulsoup4>=4.11.0
requests-html>=0.10.0
lxml_html_clean>=0.4.0
validators>=0.18.0

# Templating for emails and documents
jinja2>=3.0.0